Рефакторинг моков
